Pulled Regex class into RageUtil_Regex flies
