+++ src/Font.cpp
@@ -4,6 +4,7 @@
 
 #include "RageTextureManager.h"
 #include "RageUtil.h"
+#include "RageUtil_Regex.h"
 #include "RageLog.h"
 #include "FontManager.h"
 #include "ThemeManager.h"
